Routes Added

+++ routes/profiles.js
@@ -7,15 +7,20 @@ const ConsultantProfile = require('../models/consultantProfile');
 
 router.post('/add_seeker', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
-        const checkIfExists = await SeekerProfile.find({ _userId: req.user.id });
-        if (!checkIfExists[0]) {
-            const combinedData = {};
-            combinedData.req = req.body;
-            combinedData.req._userId = req.user.id;
-            const profileData = await new SeekerProfile(combinedData.req).save();
-            res.json(profileData)
+        if (req.user.role === 'Seeker') {
+            const checkIfExists = await SeekerProfile.find({ _userId: req.user.id });
+            console.log(checkIfExists)
+            if (!checkIfExists[0]) {
+                const combinedData = {};
+                combinedData.req = req.body;
+                combinedData.req._userId = req.user.id;
+                const profileData = await new SeekerProfile(combinedData.req).save();
+                res.json(profileData)
+            } else {
+                res.json('Edit your profile')
+            }
         } else {
-            res.json('Edit your profile')
+            res.json('You should be registered as Seeker to complete this profile')
         }
     }
     catch (e) {
@@ -25,8 +30,13 @@ router.post('/add_seeker', passport.authenticate('jwt', { session: false }), asy
 
 router.delete('/delete_seeker', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
-        const test = await SeekerProfile.deleteOne({ _userId: req.user.id });
-        res.json(test)
+        if (req.user.role === 'Seeker') {
+            const test = await SeekerProfile.deleteOne({ _userId: req.user.id });
+            res.json(test)
+        }
+        else {
+            res.json('You should be registered as Seeker to delete this profile')
+        }
     }
     catch (e) {
         throw e
@@ -35,9 +45,14 @@ router.delete('/delete_seeker', passport.authenticate('jwt', { session: false })
 
 router.patch('/edit_seeker', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
-        await SeekerProfile.updateOne({ _userId: req.user.id }, { $set: req.body });
-        const result = await SeekerProfile.find({ _userId: req.user.id });
-        res.json(result)
+        if (req.user.role === 'Seeker') {
+            await SeekerProfile.updateOne({ _userId: req.user.id }, { $set: req.body });
+            const result = await SeekerProfile.find({ _userId: req.user.id });
+            res.json(result)
+        }
+        else {
+            res.json('You should be registered as Seeker to edit this profile')
+        }
     }
     catch (e) {
         throw e
@@ -46,8 +61,13 @@ router.patch('/edit_seeker', passport.authenticate('jwt', { session: false }), a
 
 router.get('/get_seeker', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
-        const getData = await SeekerProfile.find({ _userId: req.user.id });
-        res.json(getData)
+        if (req.user.role === 'Seeker') {
+            const getData = await SeekerProfile.find({ _userId: req.user.id });
+            res.json(getData)
+        }
+        else {
+            res.json('You should be registered as Seeker to get this profile')
+        }
     }
     catch (e) {
         throw e
@@ -60,15 +80,27 @@ router.get('/get_seeker', passport.authenticate('jwt', { session: false }), asyn
 
 router.post('/add_consultant', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
-        const checkIfExists = await ConsultantProfile.find({ _userId: req.user.id });
-        if (!checkIfExists[0]) {
-            const combinedData = {};
-            combinedData.req = req.body;
-            combinedData.req._userId = req.user.id;
-            const profileData = await new ConsultantProfile(combinedData.req).save();
-            res.json(profileData)
-        } else {
-            res.json('Edit your profile')
+        if (req.user.role === 'Consultant') {
+            const checkIfExists = await ConsultantProfile.find({ _userId: req.user.id });
+            if (!checkIfExists[0]) {
+                const combinedData = {};
+                console.log(req.body)
+                combinedData.req = req.body;
+                combinedData.req._userId = req.user.id;
+                const links = [];
+                links.push({
+                    name: req.body.name,
+                    url: req.body.url
+                })
+                combinedData.req.links = links;
+                const profileData = await new ConsultantProfile(combinedData.req).save();
+                res.json(profileData)
+            } else {
+                res.json('Edit your profile')
+            }
+        }
+        else {
+            res.json('You should be registered as Consultant to add this profile')
         }
     }
     catch (e) {
@@ -78,8 +110,13 @@ router.post('/add_consultant', passport.authenticate('jwt', { session: false }),
 
 router.delete('/delete_consultant', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
-        const test = await ConsultantProfile.deleteOne({ _userId: req.user.id });
-        res.json(test)
+        if (req.user.role === 'Consultant') {
+            const test = await ConsultantProfile.deleteOne({ _userId: req.user.id });
+            res.json(test)
+        }
+        else {
+            res.json('You should be registered as Consultant to delete this profile')
+        }
     }
     catch (e) {
         throw e
@@ -88,9 +125,14 @@ router.delete('/delete_consultant', passport.authenticate('jwt', { session: fals
 
 router.patch('/edit_consultant', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
-        await ConsultantProfile.updateOne({ _userId: req.user.id }, { $set: req.body });
-        const result = await ConsultantProfile.find({ _userId: req.user.id });
-        res.json(result)
+        if (req.user.role === 'Consultant') {
+            await ConsultantProfile.updateOne({ _userId: req.user.id }, { $set: req.body });
+            const result = await ConsultantProfile.find({ _userId: req.user.id });
+            res.json(result)
+        }
+        else {
+            res.json('You should be registered as Consultant to edit this profile')
+        }
     }
     catch (e) {
         throw e
@@ -99,8 +141,13 @@ router.patch('/edit_consultant', passport.authenticate('jwt', { session: false }
 
 router.get('/get_consultant', passport.authenticate('jwt', { session: false }), async (req, res) => {
     try {
-        const getData = await ConsultantProfile.find({ _userId: req.user.id });
-        res.json(getData)
+        if (req.user.role === 'Consultant') {
+            const getData = await ConsultantProfile.find({ _userId: req.user.id });
+            res.json(getData)
+        }
+        else {
+            res.json('You should be registered as Consultant to get this profile')
+        }
     }
     catch (e) {
         throw e
